refactor : dm 패키지 정리 (dm -> conversation)
