+++ src/main/java/com/mopl/api/domain/conversation/dto/response/conversation/ConversationWith.java
@@ -1,11 +1,11 @@
-package com.mopl.api.domain.dm.dto.response.direct;
+package com.mopl.api.domain.conversation.dto.response.conversation;
 
 
 import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record DirectMessageSend (
+public record ConversationWith(
     UUID userId,
     String name,
     String profileImageUrl
