+++ src/main/java/com/mopl/api/domain/conversation/dto/response/direct/DirectMessageDto.java
@@ -1,4 +1,4 @@
-package com.mopl.api.domain.dm.dto.response.direct;
+package com.mopl.api.domain.conversation.dto.response.direct;
 
 
 import java.time.LocalDateTime;
@@ -6,13 +6,15 @@ import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record DirectMessageLastestMessage(
+public record DirectMessageDto(
+
     UUID id,
     UUID conversationId,
     LocalDateTime createdAt,
-    DirectMessageSend sender,
+    DirectMessageSend send,
     DirectMessageReceiver receiver,
     String content
+
 ) {
 
 }
