+++ src/main/java/com/mopl/api/domain/conversation/dto/response/direct/DirectMessageLastestMessage.java
@@ -1,4 +1,4 @@
-package com.mopl.api.domain.dm.dto.response.direct;
+package com.mopl.api.domain.conversation.dto.response.direct;
 
 
 import java.time.LocalDateTime;
@@ -6,15 +6,13 @@ import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record DirectMessageDto(
-
+public record DirectMessageLastestMessage(
     UUID id,
     UUID conversationId,
     LocalDateTime createdAt,
-    DirectMessageSend send,
+    DirectMessageSend sender,
     DirectMessageReceiver receiver,
     String content
-
 ) {
 
 }
