+++ src/main/java/com/mopl/api/domain/conversation/dto/response/direct/DirectMessageSend.java
@@ -1,11 +1,11 @@
-package com.mopl.api.domain.dm.dto.response.conversation;
+package com.mopl.api.domain.conversation.dto.response.direct;
 
 
 import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record ConversationWith(
+public record DirectMessageSend (
     UUID userId,
     String name,
     String profileImageUrl
