+++ src/main/java/com/mopl/api/domain/conversation/service/DirectMessageCommandServiceImpl.java
@@ -6,10 +6,10 @@ import com.mopl.api.domain.conversation.entity.DirectMessage;
 import com.mopl.api.domain.conversation.entity.QConversationParticipant;
 import com.mopl.api.domain.conversation.repository.ConversationRepository;
 import com.mopl.api.domain.conversation.repository.DirectMessageRepository;
-import com.mopl.api.domain.dm.dto.response.direct.DirectMessageDto;
-import com.mopl.api.domain.dm.dto.response.direct.DirectMessageReceiver;
-import com.mopl.api.domain.dm.dto.response.direct.DirectMessageSend;
-import com.mopl.api.domain.dm.realtime.ActiveConversationRegistry;
+import com.mopl.api.domain.conversation.dto.response.direct.DirectMessageDto;
+import com.mopl.api.domain.conversation.dto.response.direct.DirectMessageReceiver;
+import com.mopl.api.domain.conversation.dto.response.direct.DirectMessageSend;
+import com.mopl.api.domain.conversation.realtime.ActiveConversationRegistry;
 import com.mopl.api.domain.sse.SseEmitterRegistry;
 import com.mopl.api.domain.user.entity.User;
 import com.mopl.api.domain.user.repository.UserRepository;
